refactored GameTest configuration file functions. testGameConfig runs without assertion errors

+++ games4300/assign2/src/Game.h
@@ -52,7 +52,7 @@ struct PlayerConfig {
     int SR, CR;           
     int FR, FG, FB;
     int OR, OG, OB; 
-    float OT;               
+    int OT;               
     int V; 
     float S;
 };
